#10 merge some code-restructuring from old 10-* branch
