+++ lib/docuWriter/identifierSanitizer.js
@@ -5,7 +5,7 @@ var _ = require('lodash');
 /**
  * replaces forward and backward slashes with underscores
  *
- * @param inputString
+ * @param identifier
  */
 function sanitize(identifier) {
   if (_.isUndefined(identifier) || identifier === null) {
