+++ lib/docuWriter/pageNameExtractor.js
@@ -17,12 +17,12 @@ function buildDefaultPageName(urlObject) {
 }
 
 function removeEveryhtingAfterQuestionMark(href) {
-    var pos = href.indexOf('?');
-    if (pos > -1) {
-        return href.substring(0, pos);
-    } else {
-        return href;
-    }
+  var pos = href.indexOf('?');
+  if (pos > -1) {
+    return href.substring(0, pos);
+  } else {
+    return href;
+  }
 }
 
 function isValidUrlObject(urlObject) {
