+++ lib/docuWriter/util.js
@@ -1,8 +1,5 @@
 'use strict';
 
-function isDefined(value) {
-  return typeof value !== 'undefined';
-}
 
 function leadingZeros(n) {
   return ('00' + n).slice(-3);
@@ -25,7 +22,6 @@ function encodeFileName(inputString) {
 }
 
 var util = {
-  isDefined: isDefined,
   leadingZeros: leadingZeros,
   encodeFileName: encodeFileName
 };
