+++ lib/docuWriter/validationSchemas/details.json
