+++ lib/docuWriter/validationSchemas/labels.json
