+++ lib/docuWriter/validationSchemas/useCase.json
