+++ lib/scenarioo-js.js
@@ -18,90 +18,90 @@
 'use strict';
 
 var
-    _ = require('lodash'),
-    scenarioDocuWriter = require('./scenarioDocuWriter.js'),
-    scenariooJasmineReporter = require('./scenariooJasmineReporter.js');
+  _ = require('lodash'),
+  scenarioDocuWriter = require('./scenarioDocuWriter.js'),
+  scenariooJasmineReporter = require('./scenariooJasmineReporter.js');
 
 
 function describeUseCase(jasmineDescribeFunction, useCaseName, useCaseDescription, callback) {
-    if (_.isFunction(useCaseDescription)) {
-        // if useCaseDescription is omitted (since optional)
-        callback = useCaseDescription;
-        useCaseDescription = '';
-    }
+  if (_.isFunction(useCaseDescription)) {
+    // if useCaseDescription is omitted (since optional)
+    callback = useCaseDescription;
+    useCaseDescription = '';
+  }
 
-    var jasmineSuite = jasmineDescribeFunction(useCaseName, callback);
-    jasmine.scoEnv.useCases[jasmineSuite.id] = {
-        additionalDescription: useCaseDescription
-    };
-    return jasmineSuite;
+  var jasmineSuite = jasmineDescribeFunction(useCaseName, callback);
+  jasmine.scoEnv.useCases[jasmineSuite.id] = {
+    additionalDescription: useCaseDescription
+  };
+  return jasmineSuite;
 }
 
 function describeScenario(jasmineItFunction, scenarioName, scenarioDescription, callback) {
-    if (_.isFunction(scenarioDescription)) {
-        // if scenarioDescription is omitted (since optional)
-        callback = scenarioDescription;
-        scenarioDescription = '';
-    }
+  if (_.isFunction(scenarioDescription)) {
+    // if scenarioDescription is omitted (since optional)
+    callback = scenarioDescription;
+    scenarioDescription = '';
+  }
 
-    // we have to wrap our scenario with a "describe" in order to get a suite,
-    // where we can access the parent which is our current useCase
-    var jasmineSuite = describe('', function () {
-        jasmineItFunction(scenarioName, callback);
-    });
+  // we have to wrap our scenario with a "describe" in order to get a suite,
+  // where we can access the parent which is our current useCase
+  var jasmineSuite = describe('', function () {
+    jasmineItFunction(scenarioName, callback);
+  });
 
-    // use the id of the parentSuite (useCase) to store custom description
-    jasmine.scoEnv.scenarios[jasmineSuite.parentSuite.id] = {
-        additionalDescription: scenarioDescription
-    };
+  // use the id of the parentSuite (useCase) to store custom description
+  jasmine.scoEnv.scenarios[jasmineSuite.parentSuite.id] = {
+    additionalDescription: scenarioDescription
+  };
 }
 
 var scenariooJs = {
 
-    reporter: scenariooJasmineReporter,
+  reporter: scenariooJasmineReporter,
 
-    /**
-     * Expose docuWriter's "saveStep" function.
-     *
-     * Call this in your e2e test functions whenever you want scenarioo to report a step (with screen shot and metadata, etc.)
-     *
-     * @param {string} stepName
-     * @param {object} details
-     */
-    saveStep: scenarioDocuWriter.saveStep,
+  /**
+   * Expose docuWriter's "saveStep" function.
+   *
+   * Call this in your e2e test functions whenever you want scenarioo to report a step (with screen shot and metadata, etc.)
+   *
+   * @param {string} stepName
+   * @param {object} details
+   */
+  saveStep: scenarioDocuWriter.saveStep,
 
-    /**
-     * @param {string} useCaseName
-     * @param {string} [useCaseDescription] An optional description
-     * @param {function} callback
-     */
-    describeUseCase: describeUseCase.bind(undefined, describe),
+  /**
+   * @param {string} useCaseName
+   * @param {string} [useCaseDescription] An optional description
+   * @param {function} callback
+   */
+  describeUseCase: describeUseCase.bind(undefined, describe),
 
-    /**
-     * Used to run this use case exclusively (all use cases that are defined with fdescribe...)
-     *
-     * @param {string} useCaseName
-     * @param {string} [useCaseDescription] An optional description
-     * @param {function} callback
-     */
-    fdescribeUseCase: describeUseCase.bind(undefined, fdescribe),
+  /**
+   * Used to run this use case exclusively (all use cases that are defined with fdescribe...)
+   *
+   * @param {string} useCaseName
+   * @param {string} [useCaseDescription] An optional description
+   * @param {function} callback
+   */
+  fdescribeUseCase: describeUseCase.bind(undefined, fdescribe),
 
-    /**
-     *
-     * @param {string} scenarioName
-     * @param {string} [scenarioDescription] An optional description
-     * @param {function} callback
-     */
-    describeScenario: describeScenario.bind(undefined, it),
+  /**
+   *
+   * @param {string} scenarioName
+   * @param {string} [scenarioDescription] An optional description
+   * @param {function} callback
+   */
+  describeScenario: describeScenario.bind(undefined, it),
 
-    /**
-     * Used to run this use case exclusively (all scenarios that are defined with ddescribe...)
-     *
-     * @param {string} scenarioName
-     * @param {string} [scenarioDescription] An optional description
-     * @param {function} callback
-     */
-    fdescribeScenario: describeScenario.bind(undefined, fit)
+  /**
+   * Used to run this use case exclusively (all scenarios that are defined with ddescribe...)
+   *
+   * @param {string} scenarioName
+   * @param {string} [scenarioDescription] An optional description
+   * @param {function} callback
+   */
+  fdescribeScenario: describeScenario.bind(undefined, fit)
 
 };
